Optionally process movement queue in one or several background threads.
The feature is controlled by the numeric setting
[Physics]
async num threads

In case 'async num threads' > 1 and Bullet doesn't support multiple
threads, 1 async thread will be used. 0 means synchronous solver.

The PhysicsTaskScheduler is responsible for proxying requests to the collision world with proper locking, that is:
- calls to convexSweepTest and rayTest{,Internal} needs exclusive access
(Bullet singlethreaded) or shared access (Bullet multithreaded)
- other queries (contactTest, getAabb, getWorldTransform) needs shared access
- modifications (setWorldTransform, updateSingleAabb) needs exclusive
access

The main thread prepares a vector of FrameActorData, which is a snapshot
of the necessary variables before pushing it to the scheduler. In case
where it is practical, a shared atomic is used instead of a copy.

Internally, the scheduler uses 2 buffers with movement results and
standing collision maps. When called, it returns the previous buffers and
wake the background(s) thread(s) to work on the new queue.

At the end of each simulation step, the position of actors is updated
with an intermediate position. Each Actor contains a local copy of it's world
transform which is the only one accessible by the rest of the engine
(scripts for example). The physics thread access both the Actor copy and the
transform calculated by Bullet. In case of conflict between these 2
transform, the "engine version" takes precedence.
Each Actor instance ensure proper locking of its internal structure, so that
contention is practically inexistant.

Optionally update AABBs of actors and object in the background
thread(s). It is not an especially costly operation, but it needs
exclusive access to the collision world, which blocks other operations.
Since AABB needs to be updated for collision detection, one can
queue them to defer update before start of the movement solver.

The feature is controlled by the boolean setting
[Physics]
defer aabb update

Optionally compute preemptively line of sight (LOS) in the background thread(s).
After a request for LOS is made (first is synchronous, explained below),
it will be "refreshed" in the background until it is not requested
anymore and the "lineofsight keep inactive cache" number of frame has
passed.

When a getLOS call is made, it is done synchronously in the main thread,
that is the background thread is stopped during the call.
For scripts using the getLOS function, it is not really a problem
(performance wise) as these calls are very seldom. Besides, it is
unclear how to handle potentially asynchronous calls in the interpreter,
as the scripts themselves are not written to cope with that.
Other calls comes from the engine: AI packages or mechanisms such as
greetings or head tracking. AI packages generally makes such calls every
frame when triggered (which in itself is also a rare event), so they benefit
greatly from preventive refresh. Other mechanics such as greetings
should be modified to work fully asynchronously.

The feature is controlled by the numeric setting
[Physics]
lineofsight keep inactive cache